+++ seeds/store-seeds.js
@@ -1,56 +1,31 @@
-const { Item } = require('../models');
+const { User, Store, Item } = require('../models');
 
-const itemData = [
-  {
-    name: 'Potion',
-    description: 'Heals half HP',
-    effect_type: 'heal',
-    effect_amount: 50,
-    price: 50,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/potion.png'
-  },
-  {
-    name: 'Hyper Potion',
-    description: 'Heals full HP',
-    effect_type: 'heal',
-    effect_amount: 100,
-    price: 200,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/hyper-potion.png'
-  },
-  {
-    name: 'Revive',
-    description: 'Heals dead pokemon back to half HP',
-    effect_type: 'revive',
-    effect_amount: 50,
-    price: 150,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/revive.png'
-  },
-  {
-    name: 'Pokeball',
-    description: '25% chance to catch',
-    effect_type: 'catch',
-    effect_amount: 25,
-    price: 100,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/poke-ball.png'
-  },
-  {
-    name: 'Ultra Ball',
-    description: '65% chance to catch',
-    effect_type: 'catch',
-    effect_amount: 65,
-    price: 300,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/ultra-ball.png'
-  },
-  {
-    name: 'Master Ball',
-    description: '100% chance to catch',
-    effect_type: 'catch',
-    effect_amount: 100,
-    price: 500,
-    image: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/items/master-ball.png'
-  },
-];
+const userStoreData = [];
 
-const seedItems = () => Item.bulkCreate(itemData);
 
-module.exports = seedItems;
+const seedStores = async () => {
+    try {
+        // get all users
+        const userData = await User.findAll();
+        // get all items
+        const itemData = await Item.findAll();
+
+        // for every user
+        for (let i = 0; i < userData.length; i++) {
+            // for every item
+            for (let x = 0; x < itemData.length; x++) {
+                // add user + items in userStoreData
+                userStoreData.push({
+                    count: 3,
+                    user_id: userData[i].id,
+                    item_id: itemData[x].id
+                })
+            }
+        }
+        await Store.bulkCreate(userStoreData);
+    } catch (error) {
+        console.error('Error seeding wallets:', error);
+    }
+};
+
+module.exports = seedStores;
